VACMS-4586: Move GTM tests to Cypress
