+++ tests/behavioral/cypress/integration/common/i_am_an_anonymous_user.js
@@ -1,0 +1,5 @@
+import { Given } from "cypress-cucumber-preprocessor/steps";
+
+Given(`I am an anonymous user`, () => {
+  /** This step definition intentionally left blank. */
+});
