+++ tests/behavioral/cypress/integration/google_tag_manager/the_gtm_data_layer_value_for_key_should_be_unset.js
@@ -1,0 +1,8 @@
+import { Then } from "cypress-cucumber-preprocessor/steps";
+
+Then(`the GTM data layer value for {string} should be unset`, (key) => {
+  cy.window().then((window) => {
+    const actual = window.drupalSettings.gtm_data[key];
+    cy.wrap(actual).should('eq', undefined);
+  });
+});
